requirements.txt
cyvcf2==0.30.12
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
PySimpleGUI==4.55.1
scikit-learn==1.0.1
umap-learn==0.5.2